Switch from field injection to constructor injection for beans (#230)
## Motivation

Field injection is not recommended in Spring (for example, IntelliJ shows a warning). There are a couple of articles and discussions showing the advantages of Constructor-based dependency injection and Setter-based dependency injection. For this project, I recommend switching from Field-based dependency injection to Constructor-based injection. It should bring better testability, immutability, and show problems in design.

## Change

Initialization of all fields which were injected with the `@Autowired` annotation was moved to constructors.