Fix URL for example in README.md (#475)
* Fix URL for example in README.md

* Update README.md

### Motivation

Attempting to use this section of the documentation, the intent of the initial documented link to localhost was clear but the implementation failed to include the port number, causing the link to be confusing at best. 

### Modifications

Added a port number to the URL in the example. 

### Verifying this change

No Checks should be required, as this is a documentation update, with valid Markdown.